+++ data-transfer-object/src/main/java/org/grnet/status/dtos/tenant/TenantRequestDto.java
@@ -38,5 +38,4 @@ public class TenantRequestDto {
     @JsonProperty("metadata")
     @Valid
     public TenantMetadata metadata;
-
 }