+++ entity/src/main/resources/db/migration/tables/tenant/V1.10__update_t_tenant_v2.sql
@@ -1,0 +1,8 @@
+-- ------------------------------------------------
+-- Version: V1.10
+-- Description: Update t_tenant table to add status column
+-- ------------------------------------------------
+ALTER TABLE t_Tenant
+ADD COLUMN status JSONB DEFAULT '{}'::jsonb;
+ALTER TABLE t_Tenant
+ALTER COLUMN status SET NOT NULL;