+++ enum/src/main/java/org/grnet/status/enums/EventName.java
@@ -1,0 +1,6 @@
+package org.grnet.status.enums;
+
+public enum EventName {
+        INIT_AMS,
+        INIT_MONGO
+}