+++ enum/src/main/java/org/grnet/status/enums/EventStatus.java
@@ -1,0 +1,8 @@
+package org.grnet.status.enums;
+
+public enum EventStatus {
+    IN_PROGRESS,
+    COMPLETED,
+    FAILED
+}
+